Money value stored as int64 instead of float64

+++ aggregators/aggregator_func.go
@@ -4,8 +4,8 @@ import (
 	"github.com/apoydence/ledger/transaction"
 )
 
-type AggregatorFunc func([]*transaction.Account) float64
+type AggregatorFunc func([]*transaction.Account) int64
 
-func (a AggregatorFunc) Aggregate(accounts []*transaction.Account) float64 {
+func (a AggregatorFunc) Aggregate(accounts []*transaction.Account) int64 {
 	return a(accounts)
 }
