+++ aggregators/store_test.go
@@ -12,7 +12,7 @@ var _ = Describe("Store", func() {
 	var someAgg aggregators.AggregatorFunc
 
 	BeforeEach(func() {
-		someAgg = aggregators.AggregatorFunc(func([]*transaction.Account) float64 {
+		someAgg = aggregators.AggregatorFunc(func([]*transaction.Account) int64 {
 			return 0
 		})
 	})
